Isolate `winston` code in separate file
