+++ src/util/logger/index.ts
@@ -1,0 +1,11 @@
+/**
+ * Copyright (c) 2024-present, NDLA.
+ *
+ * This source code is licensed under the GPLv3 license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ */
+
+import log from "./logger";
+export { default as log } from "./logger";
+export default log;
